requirements.txt
django>=4.2,<5.0
djangorestframework
requests
pyyaml
django-cors-headers
